fix: Don't auto retry SSLError
We were observing that SSLErrors would occur for some users when
wait_for was called due to the websockets library auto retrying
all OSErrors and SSLError inherititing from OSError. These errors
are not actually recoverable in most cases and either indicate
a configuration issue in the server or the client. Instead we should
abort out as quickly as possible and let the user know that there
is an issue with their certificates.

As to why some users have SSL errors I am unsure and that requires
a further investigation, but this should at least give them some
immediate feedback.